pyproject.toml
[tool.poetry]
name = "workshop-regex-pdf"
version = "0.1.0"
description = ""
authors = ["Marcos Soares <marcosjmsoares@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
camelot-py = "^0.11.0"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.3"
ghostscript = "^0.7"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.32"
unidecode = "^1.3.8"
pypdf2 = "^3.0.1"
black = "^24.8.0"
isort = "^5.13.2"
schedule = "^1.2.2"
pytz = "^2024.2"
boto3 = "^1.35.29"
python-dotenv = "^1.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
